mod to AvoidObstacle behavior to induce allstop when collision unavoidable (#27)
Mod to AvoidObstacle behavior to induce allstop when collision with
obstacle is unavoidable. Mod to PDMap and IvPFunction classes to support
checks for this. Prior to this mod, the behavior would essentially
produce not IvP function since all choices were ranked 0. Which had the
effect of essentially turning off the obstacle avoidance behavior when a
collision became unavoidable. Not it will induce a helm all-stop by
posting a behavior error. Which is better than continuing ahead at
speed. This situation was detected in a simulation where the starting
vehicle position, heading and speed was initially in a relation to an
obstacle immediately in front of the vessel.