Mod to IvPBehavior to only invoke complete for stale contact behavior… (#57)
Mod to IvPBehavior to only invoke setComplete for stale contact behavior
if it was dynamically spawned. This fixes a problem with the recent
change to allow contact behaviors to self-complete when/if the contact
disappears from the helm ledger (presumably due to a period where no
NODE_REPORTS have been received for a contact). This was a problem for
static contact behaviors with their initial contact name set to
something like "to-be-set".