AvoidObstacle bhv bugfix, ownship starts in obstacle (#60)
Addresses bug that occurred when/if ownship starts IN a polygon representing an obstacle. Would result in helm crash originating the BHV_AvoidObstacleV24 behavior. Also changed the behavior to post a BHV_ERROR resulting in a helm all-stop if for some reason ownship has been detected to breach the obstacle polygon. Bug largely not produceable under normal circumstance of the vehicle approaching an obstacle rather than starting within it. Also updated the _Info files for pMarineViewer and pNodeReporter.